Implement quantile aggregate functions.
